fixup: fix `Message` format
Fixes the `MessageFormat` to include the request/event code byte.

Unifies `FuncId` and `FunctionMode` into one type, they are semantically
identical.

TODO: merge `Message::func_id` and `Message::code` into `MessageCode`.
