+++ src/func_id.rs
@@ -9,10 +9,10 @@ const ESCROW: u8 = 0b0011;
 const RESERVED: u8 = 0xff;
 const STATUS_SHIFT: u8 = 12;
 
-/// Represents the function mode of the JCM device.
+/// Represents the function ID of the JCM device.
 #[repr(u8)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum FunctionMode {
+pub enum FuncId {
     /// Common (entire device).
     Common = COMMON,
     /// Acceptor function.
@@ -25,13 +25,13 @@ pub enum FunctionMode {
     Reserved = RESERVED,
 }
 
-impl FunctionMode {
-    /// Creates a new [FunctionMode].
+impl FuncId {
+    /// Creates a new [FuncId].
     pub const fn new() -> Self {
         Self::Common
     }
 
-    /// Infallible conversion from a [`u8`] into a [FunctionMode].
+    /// Infallible conversion from a [`u8`] into a [FuncId].
     pub const fn from_u8(val: u8) -> Self {
         match val {
             COMMON => Self::Common,
@@ -42,34 +42,34 @@ impl FunctionMode {
         }
     }
 
-    /// Infallible conversion from a [`u16`] into a [FunctionMode].
+    /// Infallible conversion from a [`u16`] into a [FuncId].
     pub const fn from_u16(val: u16) -> Self {
         Self::from_u8((val >> STATUS_SHIFT) as u8)
     }
 
-    /// Gets the length of the [FunctionMode].
+    /// Gets the length of the [FuncId].
     pub const fn len() -> usize {
         mem::size_of::<u8>()
     }
 
-    /// Gets whether the [FunctionMode] contains a reserved variant.
+    /// Gets whether the [FuncId] contains a reserved variant.
     pub const fn is_empty(&self) -> bool {
         matches!(self, Self::Reserved)
     }
 }
 
-impl TryFrom<u8> for FunctionMode {
+impl TryFrom<u8> for FuncId {
     type Error = Error;
 
     fn try_from(val: u8) -> Result<Self> {
         match Self::from_u8(val) {
-            Self::Reserved => Err(Error::InvalidFunctionMode(val)),
+            Self::Reserved => Err(Error::InvalidFuncId(val)),
             v => Ok(v),
         }
     }
 }
 
-impl TryFrom<u16> for FunctionMode {
+impl TryFrom<u16> for FuncId {
     type Error = Error;
 
     fn try_from(val: u16) -> Result<Self> {
@@ -77,31 +77,31 @@ impl TryFrom<u16> for FunctionMode {
     }
 }
 
-impl From<FunctionMode> for u16 {
-    fn from(val: FunctionMode) -> Self {
+impl From<FuncId> for u16 {
+    fn from(val: FuncId) -> Self {
         (val as u16) << STATUS_SHIFT
     }
 }
 
-impl From<&FunctionMode> for &'static str {
-    fn from(val: &FunctionMode) -> Self {
+impl From<&FuncId> for &'static str {
+    fn from(val: &FuncId) -> Self {
         match val {
-            FunctionMode::Common => "common",
-            FunctionMode::Acceptor => "acceptor",
-            FunctionMode::Recycler => "recycler",
-            FunctionMode::Escrow => "escrow",
-            FunctionMode::Reserved => "reserved",
+            FuncId::Common => "common",
+            FuncId::Acceptor => "acceptor",
+            FuncId::Recycler => "recycler",
+            FuncId::Escrow => "escrow",
+            FuncId::Reserved => "reserved",
         }
     }
 }
 
-impl From<FunctionMode> for &'static str {
-    fn from(val: FunctionMode) -> Self {
+impl From<FuncId> for &'static str {
+    fn from(val: FuncId) -> Self {
         (&val).into()
     }
 }
 
-impl fmt::Display for FunctionMode {
+impl fmt::Display for FuncId {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, r#""{}""#, <&str>::from(self))
     }
@@ -115,20 +115,20 @@ mod tests {
     fn test_function_mode() {
         let raw_denom = [COMMON, ACCEPTOR, RECYCLER, ESCROW];
         let expected = [
-            FunctionMode::Common,
-            FunctionMode::Acceptor,
-            FunctionMode::Recycler,
-            FunctionMode::Escrow,
+            FuncId::Common,
+            FuncId::Acceptor,
+            FuncId::Recycler,
+            FuncId::Escrow,
         ];
 
         for (raw, exp) in raw_denom.into_iter().zip(expected.into_iter()) {
-            assert_eq!(FunctionMode::try_from(raw), Ok(exp));
-            assert_eq!(FunctionMode::from_u8(raw), exp);
+            assert_eq!(FuncId::try_from(raw), Ok(exp));
+            assert_eq!(FuncId::from_u8(raw), exp);
         }
 
         for stat in (0..=255u8).filter(|s| !raw_denom.iter().any(|d| d == s)) {
-            assert!(FunctionMode::try_from(stat).is_err());
-            assert_eq!(FunctionMode::from_u8(stat), FunctionMode::Reserved);
+            assert!(FuncId::try_from(stat).is_err());
+            assert_eq!(FuncId::from_u8(stat), FuncId::Reserved);
         }
     }
 }
