+++ src/lib.rs
@@ -3,6 +3,7 @@ mod denomination;
 mod device_status;
 mod error;
 mod failure_code;
+mod func_id;
 mod message;
 mod status_code;
 
@@ -11,5 +12,6 @@ pub use denomination::*;
 pub use device_status::*;
 pub use error::*;
 pub use failure_code::*;
+pub use func_id::*;
 pub use message::*;
 pub use status_code::*;
